# api/tsclient/src/models/Note.ts
/* tslint:disable */
/* eslint-disable */
/**
 * OpenAPI
 * Notes API for Learning
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface Note
 */
export interface Note {
    /**
     * 
     * @type {number}
     * @memberof Note
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof Note
     */
    message: string;
    /**
     * 
     * @type {Array<string>}
     * @memberof Note
     */
    tags?: Array<string>;
}

/**
 * Check if a given object implements the Note interface.
 */
export function instanceOfNote(value: object): value is Note {
    if (!('message' in value) || value['message'] === undefined) return false;
    return true;
}

export function NoteFromJSON(json: any): Note {
    return NoteFromJSONTyped(json, false);
}

export function NoteFromJSONTyped(json: any, ignoreDiscriminator: boolean): Note {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'message': json['message'],
        'tags': json['tags'] == null ? undefined : json['tags'],
    };
}

export function NoteToJSON(json: any): Note {
    return NoteToJSONTyped(json, false);
}

export function NoteToJSONTyped(value?: Note | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'message': value['message'],
        'tags': value['tags'],
    };
}